Change default command and add debugging tip.
On a brand new OS X installation of Yosemite on my MacBook Air, the script needs to be run with `bash`, and subsequent runs with `bash` work just fine.

For some reason, on my 18F MacBook Pro, which already had a dev environment, the RVM sourcing line fails when run with `bash`. Running it with `sh` fixes it.