Fix RVM sourcing issue.
There seems to be a bug with running this script if a version of RVM prior to 1.26.10 was already installed on the machine. The bug is that the RVM sourcing line (`. ~/.rvm/scripts/rvm`) will fail when the script is run with `bash`.

It turns out that sourcing within the script is only needed when first installing RVM, so by moving the sourcing line to the conditional, this issue gets fixed.

In addition, reloading of RVM after upgrading from RVM 1.26.9 to 1.26.10 will fail (this is an RVM bug), so I've added the auto reload flag to `.rvmrc`.