Install Bundler globally
RVM stopped shipping with Bundler, which means that if you clone
a repo that uses a `.ruby-gemset`, or if you install a new version
of Ruby, or update to a new version of RVM, Bundler won't be
automatically available, and `bundle install` will fail.

Beginners might not know they can `gem install bundler` to fix the
issue, and even if you did know, having to type that each time is
annoying. Since Bundler is widely used, it makes sense to install
it in the global gemset so that it is available at any time.

RVM provides the  `--with-gems` flag that can be used to install
bundler when installing RVM for the first time, and when updating
RVM. However, there is a bug where using the flag when installing
for the first time will prevent Ruby from being installed.
Therefore, I only enabled the flag for the upgrade scenario.

I also had to fix the append_to_file function to look for exact
string matches. Otherwise, it would treat "rubygems-bundler" as a
match for "bundler" and won't add it to the global gemset.