Skip cf-cli and brew doctor failures
The latest cf-cli release is resulting in a SHA1 mismatch error.
This change allows the script to continue running if cf-cli failed to
install, and asks the user to report the issue on GitHub.

Also, after installing Python via pyenv, `brew doctor` will report
some warnings about the Python setup. We need to be able to allow
the script to continue running after any errors or warning reported
by `brew doctor`.