Use `brew bundle` instead of custom functions
**Why**:
Homebrew-bundle does what the custom functions used to do. It also allows us to
use a Brewfile, which makes it easier to see in one place what gets installed,
and makes it easier to customize that list.

As part of moving to a Brewfile, I also removed the directives to restart
Redis, Postgres, and MySQL. Given that the script installs Homebrew services,
managing services manually is very easy, and the user can choose when to start
them.