Specify Brewfile path
**Why**: Due to the new way of installing the script via curl and
piping to `sh`, it doesn’t see the `Brewfile`, so we must explicitly
specify its path.

Also, allow `brew bundle` to fail and continue because we don't want
to exit the script if the reason for failure was that a Mac app that
Cask is trying to install was already installed on the machine.